Feat: 이미지 섹션 컴포넌트로 추상화 #55
