+++ components/ui/ModalRefactor/AddChatRoomModal/index.tsx
@@ -4,50 +4,26 @@
 /* eslint-disable jsx-a11y/no-static-element-interactions */
 
 import { useState } from 'react';
-import Image from 'next/image';
 import { ModalProps } from '@/types/types';
 import useInput from '@/hooks/common/useInput';
 import XIcon from '@/public/X.svg';
 import ArrowLeftIcon from '@/public/arrow-left.svg';
-import CameraIcon from '@/public/Camera.svg';
 import Modal from '../../Modal';
 import TitleInput from './TitleInput';
 import DescriptionInput from './DescriptionInput';
 import HashTagInput from './HashTagInput';
 import Button from '../../Button';
 import Divider from '../../Divider';
-
-const FirstImageList = [
-  '/images/AddChatModal/default2.webp',
-  '/images/AddChatModal/default3.webp',
-  '/images/AddChatModal/default4.webp',
-  '/images/AddChatModal/default5.webp',
-  '/images/AddChatModal/default6.webp',
-  '/images/AddChatModal/default7.webp',
-  '/images/AddChatModal/default8.webp',
-];
-const SecondImageList = [
-  '/images/AddChatModal/default9.webp',
-  '/images/AddChatModal/default10.webp',
-  '/images/AddChatModal/default11.webp',
-  '/images/AddChatModal/default12.webp',
-];
+import ImageDisplay from './ImageDisplay';
+import ImageList from './ImageList';
 
 export default function AddChatRoomModal({ open, onClose }: ModalProps) {
   const [title, onChangeTitle] = useInput('');
   const [description, onChangeDescription] = useInput('');
   const [tag, onChangeTag, resetTag] = useInput('');
   const [tagList, setTagList] = useState<string[]>([]);
-  const [imaegList, setImageList] = useState<string[]>(FirstImageList);
   const [image, setImage] = useState('/images/AddChatModal/default2.webp');
   const [isChecked, setIsChecked] = useState('');
-
-  const onNetxImageList = () => {
-    setImageList(SecondImageList);
-  };
-  const onPrevImageList = () => {
-    setImageList(FirstImageList);
-  };
 
   return (
     <Modal open={open} onClose={onClose}>
@@ -68,89 +44,22 @@ export default function AddChatRoomModal({ open, onClose }: ModalProps) {
           <div className="flex flex-col order-2 gap-3 tablet:order-1">
             <TitleInput title={title} onChangeTitle={onChangeTitle} />
             <Divider type="horizontal" />
-
-            <div className="flex flex-col gap-3">
-              <DescriptionInput
-                description={description}
-                onChangeDescription={onChangeDescription}
-              />
-              <HashTagInput
-                onChangeTag={onChangeTag}
-                reset={resetTag}
-                setTagList={setTagList}
-                tag={tag}
-                tagList={tagList}
-              />
-            </div>
+            <DescriptionInput
+              description={description}
+              onChangeDescription={onChangeDescription}
+            />
+            <HashTagInput
+              onChangeTag={onChangeTag}
+              reset={resetTag}
+              setTagList={setTagList}
+              tag={tag}
+              tagList={tagList}
+            />
           </div>
 
           <div className="flex items-center justify-between order-1 gap-4 tablet:order-2">
-            <div className="relative flex-1 h-44 tablet:h-60">
-              <Image
-                src={image}
-                alt=""
-                fill
-                sizes="100vw"
-                priority
-                className="rounded-[10px] tablet:rounded-none"
-              />
-              <button
-                onClick={() => console.log('d')}
-                type="button"
-                className="absolute bottom-0 right-0 flex items-center justify-center w-12 h-12 mb-2 mr-2 rounded-full bg-grey-200 tablet:hidden"
-              >
-                <CameraIcon />
-              </button>
-            </div>
-            <div className="flex-col flex-1 hidden gap-5 tablet:flex">
-              <div className="flex justify-between">
-                <span className="text-grey-600 body3">커버선택</span>
-                <div className="flex gap-2">
-                  <span className="caption2 text-grey-500">1 / 2</span>
-                  <div>
-                    <button type="button" onClick={onPrevImageList}>
-                      이전
-                    </button>
-                    <button type="button" onClick={onNetxImageList}>
-                      다음
-                    </button>
-                  </div>
-                </div>
-              </div>
-              <ul className="grid grid-cols-4 gap-2 ">
-                <li>
-                  <label
-                    className="flex flex-col items-center justify-center w-full h-full cursor-pointer bg-grey-200"
-                    htmlFor="imageUpload"
-                  >
-                    <input
-                      type="file"
-                      className="hidden"
-                      id="imageUpload"
-                      accept="image/jpg,impge/png,image/jpeg,image/gif,image/webp"
-                    />
-                    <CameraIcon />
-                    <span className="capion1 text-grey-400">사진추가</span>
-                  </label>
-                </li>
-                {imaegList.map((img) => (
-                  // eslint-disable-next-line jsx-a11y/no-noninteractive-element-interactions
-                  <li
-                    key={img}
-                    onClick={() => setImage(img)}
-                    className="cursor-pointer "
-                  >
-                    <Image
-                      src={img}
-                      alt=""
-                      width={120}
-                      height={90}
-                      className="h-24 duration-150 w-30 hover:scale-105"
-                    />
-                  </li>
-                ))}
-              </ul>
-            </div>
+            <ImageDisplay image={image} />
+            <ImageList setImage={setImage} />
           </div>
 
           <div className="flex flex-col order-3 gap-6">
